more testing friendly
